test fault under editing

+++ kern/tests/test_working_set.c
@@ -9,21 +9,21 @@
 #include <kern/proc/user_environment.h>
 #include <kern/mem/working_set_manager.h>
 
-//2020
-int sys_check_LRU_lists(uint32* active_list_content, uint32* second_list_content, int actual_active_list_size, int actual_second_list_size)
+// 2020
+int sys_check_LRU_lists(uint32 *active_list_content, uint32 *second_list_content, int actual_active_list_size, int actual_second_list_size)
 {
-	struct Env* cur_env = get_cpu_proc();
+	struct Env *cur_env = get_cpu_proc();
 	assert(cur_env != NULL);
 	cprintf("CURRENT WS CONTENT BEFORE CHECKING:\n");
 	env_page_ws_print(cur_env);
 
-	struct Env* env = cur_env;
+	struct Env *env = cur_env;
 	int active_list_validation = 1;
 	int second_list_validation = 1;
-	struct WorkingSetElement* ptr_WS_element;
+	struct WorkingSetElement *ptr_WS_element;
 
-	//1- Check active list content if not null
-	if(active_list_content != NULL)
+	// 1- Check active list content if not null
+	if (active_list_content != NULL)
 	{
 		int idx_active_list = 0;
 		LIST_FOREACH(ptr_WS_element, &(env->ActiveList))
@@ -35,15 +35,14 @@ int sys_check_LRU_lists(uint32* active_list_content, uint32* second_list_content
 			}
 			idx_active_list++;
 		}
-		if(LIST_SIZE(&env->ActiveList) != actual_active_list_size)
+		if (LIST_SIZE(&env->ActiveList) != actual_active_list_size)
 		{
 			active_list_validation = 0;
-
 		}
 	}
 
-	//2- Check second chance list content if not null
-	if(second_list_content != NULL)
+	// 2- Check second chance list content if not null
+	if (second_list_content != NULL)
 	{
 		int idx_second_list = 0;
 		LIST_FOREACH(ptr_WS_element, &(env->SecondList))
@@ -55,25 +54,24 @@ int sys_check_LRU_lists(uint32* active_list_content, uint32* second_list_content
 			}
 			idx_second_list++;
 		}
-		if(LIST_SIZE(&env->SecondList) != actual_second_list_size)
+		if (LIST_SIZE(&env->SecondList) != actual_second_list_size)
 			second_list_validation = 0;
 	}
-	return active_list_validation&second_list_validation;
+	return active_list_validation & second_list_validation;
 }
 
-
-//2020
-int sys_check_LRU_lists_free(uint32* list_content, int list_size)
+// 2020
+int sys_check_LRU_lists_free(uint32 *list_content, int list_size)
 {
-	struct Env* cur_env = get_cpu_proc();
+	struct Env *cur_env = get_cpu_proc();
 	assert(cur_env != NULL);
-	struct Env* env = cur_env;
+	struct Env *env = cur_env;
 	int list_validation_count = 0;
-	struct WorkingSetElement* ptr_WS_element;
+	struct WorkingSetElement *ptr_WS_element;
 
 	LIST_FOREACH(ptr_WS_element, &(env->ActiveList))
 	{
-		for(int var = 0; var < list_size; var++)
+		for (int var = 0; var < list_size; var++)
 		{
 			if (ROUNDDOWN(ptr_WS_element->virtual_address, PAGE_SIZE) == ROUNDDOWN(list_content[var], PAGE_SIZE))
 			{
@@ -81,14 +79,13 @@ int sys_check_LRU_lists_free(uint32* list_content, int list_size)
 				break;
 			}
 		}
-		if(list_validation_count > 0)
+		if (list_validation_count > 0)
 			return list_validation_count;
 	}
-
 
 	LIST_FOREACH(ptr_WS_element, &(env->SecondList))
 	{
-		for(int var = 0; var < list_size; var++)
+		for (int var = 0; var < list_size; var++)
 		{
 			if (ROUNDDOWN(ptr_WS_element->virtual_address, PAGE_SIZE) == ROUNDDOWN(list_content[var], PAGE_SIZE))
 			{
@@ -96,41 +93,39 @@ int sys_check_LRU_lists_free(uint32* list_content, int list_size)
 				break;
 			}
 		}
-		if(list_validation_count > 0)
+		if (list_validation_count > 0)
 			return list_validation_count;
-
 	}
-
 
 	return list_validation_count;
 }
 
-//2023
+// 2023
 /*chk_status:
  * = 0: check entire list (order is not important)
  * = 1: check entire list (order is important)
  * = 2: check only the existence of the given set of elements
  * = 3: check only the NOT existence of the given set of elements
  */
-int sys_check_WS_list(uint32* WS_list_content, int actual_WS_list_size, uint32 last_WS_element_content, bool chk_status)
+int sys_check_WS_list(uint32 *WS_list_content, int actual_WS_list_size, uint32 last_WS_element_content, bool chk_status)
 {
 #if USE_KHEAP
-//	cprintf("CURRENT WS CONTENT BEFORE CHECKING:\n");
-//	env_page_ws_print(get_cpu_proc());
-	struct Env* cur_env = get_cpu_proc();
+	//	cprintf("CURRENT WS CONTENT BEFORE CHECKING:\n");
+	//	env_page_ws_print(get_cpu_proc());
+	struct Env *cur_env = get_cpu_proc();
 	assert(cur_env != NULL);
-	struct Env* env = cur_env;
+	struct Env *env = cur_env;
 	int WS_list_validation = 1;
-	struct WorkingSetElement* ptr_WS_element;
+	struct WorkingSetElement *ptr_WS_element;
 
 	if (chk_status == 0 || chk_status == 1)
 	{
-		if(LIST_SIZE(&(env->page_WS_list)) != actual_WS_list_size)
+		if (LIST_SIZE(&(env->page_WS_list)) != actual_WS_list_size)
 		{
 			return WS_list_validation = 0;
 		}
 	}
-	//if it's required to check the last_WS_element
+	// if it's required to check the last_WS_element
 	if (last_WS_element_content != 0)
 	{
 		if (ROUNDDOWN(env->page_last_WS_element->virtual_address, PAGE_SIZE) != ROUNDDOWN(last_WS_element_content, PAGE_SIZE))
@@ -138,7 +133,7 @@ int sys_check_WS_list(uint32* WS_list_content, int actual_WS_list_size, uint32 l
 			return WS_list_validation = 0;
 		}
 	}
-	//if the order of the content is important to check
+	// if the order of the content is important to check
 	if (chk_status == 1)
 	{
 		//		int idx_WS_list = 0;
@@ -153,30 +148,31 @@ int sys_check_WS_list(uint32* WS_list_content, int actual_WS_list_size, uint32 l
 		//		}
 		int idx_WS_list = 0;
 
-		//Search for the correct index of the current WS element (if any)
+		// Search for the correct index of the current WS element (if any)
 		if (last_WS_element_content)
 		{
 			for (int i = 0; i < actual_WS_list_size; ++i)
 			{
 				if (ROUNDDOWN(WS_list_content[i], PAGE_SIZE) == ROUNDDOWN(last_WS_element_content, PAGE_SIZE))
 				{
-					idx_WS_list = i ;
+					idx_WS_list = i;
 					break;
 				}
 			}
 		}
-		//cprintf("index of last WS element = %d\n",idx_WS_list);
-		//Check the expected content starting from last WS element (if any)
+		// cprintf("index of last WS element = %d\n",idx_WS_list);
+		// Check the expected content starting from last WS element (if any)
 		if (env->page_last_WS_element)
 			ptr_WS_element = env->page_last_WS_element;
 		else
 			ptr_WS_element = LIST_FIRST(&(env->page_WS_list));
 
-		//cprintf("comparison start from va = %x\n",ptr_WS_element->virtual_address);
+		// cprintf("comparison start from va = %x\n",ptr_WS_element->virtual_address);
 
 		for (int i = 0; i < actual_WS_list_size; ++i)
 		{
-			if (ROUNDDOWN(ptr_WS_element->virtual_address, PAGE_SIZE) != ROUNDDOWN(WS_list_content[idx_WS_list], PAGE_SIZE))
+
+			if (ROUNDDOWN(ptr_WS_element->virtual_address, PAGE_SIZE) != (uint32)ROUNDDOWN(WS_list_content[idx_WS_list], PAGE_SIZE))
 			{
 				WS_list_validation = 0;
 				break;
@@ -207,7 +203,7 @@ int sys_check_WS_list(uint32* WS_list_content, int actual_WS_list_size, uint32 l
 			}
 		}
 	}
-	//Check NON-EXITENCE of the Given Addresses
+	// Check NON-EXITENCE of the Given Addresses
 	else if (chk_status == 3)
 	{
 		for (int idx_expected_list = 0; idx_expected_list < actual_WS_list_size; ++idx_expected_list)
